requirements.txt
# Core dependencies for Python 3.12 (October 2025)
anthropic==0.39.0
openai==1.54.3
google-generativeai==0.8.3

# Data processing
numpy==2.1.2
pandas==2.2.3
scipy==1.14.1

# NLP and metrics
nltk==3.9.1
spacy==3.8.2
sentence-transformers==3.2.1
transformers==4.46.2

# Analysis and visualization
jupyter==1.1.1
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Statistics
statsmodels==0.14.4
pingouin==0.5.5

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.3
pytest-cov==6.0.0

# Utilities
tqdm==4.66.5
requests==2.32.3